Updating src/pieces/pawn.
This commit updates the pawn class to use the slide direction
module for legal moves and account for the differences in
black and white. Before both color pieces would be trying to move
in the same direction. This is fixed. The pawn class also behaves
that the piece has not moved when it is created rather then
checking the location of the piece. The location was also taken
off the class to follow the changes made to the piece class.