Creates src/lib/getAllMoves.js
This commit creates the get all moves module which returns all
the legal oves for both players along with the location of both
players kings. This infromation is going to be needed in order
to determine who is in check and if making a move would put
a player into check.